pyproject.toml
[tool.mypy]
ignore_missing_imports = true
plugins = ["pydantic.mypy"]

[tool.poetry]
name = "cytomine-cbir"
version = "0.1.0"
description = ""
authors = ["Ba Thien Le <ba.le@uliege.be>"]
license = "Apache 2.0"
readme = "README.md"
repository = "https://github.com/Cytomine-ULiege/Cytomine-cbir"
packages = [{include = "cbir"}]

[tool.poetry.dependencies]
python = "^3.9"
cbir-tfe = {git = "https://github.com/Cytomine-ULiege/cbir-tfe.git", rev = "master"}
faiss-gpu = "^1.7.2"
fastapi = "^0.104.1"
numpy = "^1.26.2"
pillow = "^10.1.0"
pydantic-settings = "^2.0.3"
python-multipart = "^0.0.6"
redis = "3.5.3"
torch = {version = "2.1.0+cu118", source = "torch-gpu"}
torchvision = {version = "^0.16.0+cu118", source = "torch-gpu"}
uvicorn = {extras = ["standard"], version = "^0.24.0.post1"}

[tool.poetry.group.dev.dependencies]
black = "^23.10.1"
codespell = "^2.2.6"
coverage = "^7.3.2"
isort = "^5.12.0"
pylint = "^3.0.2"
mypy = {extras = ["reports"], version = "^1.7.0"}

[tool.poetry.group.test.dependencies]
pytest = "^7.4.3"
httpx = "^0.25.1"
dvc = {extras = ["gdrive"], version = "^3.29.0"}

[[tool.poetry.source]]
name = "torch-gpu"
url = "https://download.pytorch.org/whl/cu118"
priority = "supplemental"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
